requirements.txt
torch>=2.1.0
transformers>=4.38.0
numpy>=1.24.0
PyYAML>=6.0.0
tqdm>=4.66.0
gradio>=4.44.0
